backend-django/task_management/requirements.txt
asgiref==3.5.0
distlib==0.3.6
Django==4.0.1
django-binary-database-files==1.0.10
django-cors-headers==3.14.0
djangorestframework==3.14.0
filelock==3.10.0
future==0.18.3
Pillow==9.0.0
platformdirs==3.1.1
psycopg2==2.9.3
pytz==2021.3
serializers==0.2.4
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
virtualenv==20.21.0
